requirements.txt
fastapi
uvicorn[standard]
gunicorn
supabase-py
python-telegram-bot
python-dotenv
redis
rq
psutil 